refactor: 패키지 및 화면 구조 변경 - base 파일 추가 - navigation 제거 - 화면 추가
issue:#15
