+++ app/src/main/AndroidManifest.xml
@@ -3,8 +3,10 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <uses-feature android:name="android.hardware.camera.any" />
+
     <uses-permission android:name="android.permission.CAMERA" />
-    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"
+    <uses-permission
+        android:name="android.permission.WRITE_EXTERNAL_STORAGE"
         android:maxSdkVersion="28" />
 
     <application
@@ -18,21 +20,22 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Snapsketch"
         tools:targetApi="31">
-
+        <activity
+            android:name=".presentation.ui.camera.CameraActivity"
+            android:exported="false" />
         <activity
             android:name=".presentation.ui.splash.SplashActivity"
             android:exported="true"
             android:theme="@style/SplashTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
         <activity
             android:name=".presentation.ui.main.MainActivity"
             android:exported="false" />
-
         <activity
             android:name=".presentation.ui.coloring.ColoringActivity"
             android:exported="false" />
