+++ app/src/main/java/com/project/snapsketch/data/local/LocalDataSource.kt
@@ -1,0 +1,4 @@
+package com.project.snapsketch.data.local
+
+interface LocalDataSource {
+}
