+++ app/src/main/java/com/project/snapsketch/data/local/LocalDataSourceImpl.kt
@@ -1,0 +1,4 @@
+package com.project.snapsketch.data.local
+
+class LocalDataSourceImpl {
+}
