+++ app/src/main/java/com/project/snapsketch/data/repository/ImageRepositoryImpl.kt
@@ -1,0 +1,4 @@
+package com.project.snapsketch.data.repository
+
+class ImageRepositoryImpl {
+}
