+++ app/src/main/java/com/project/snapsketch/domain/opencv/EdgeDetectionProcessor.kt
@@ -1,0 +1,4 @@
+package com.project.snapsketch.domain.opencv
+
+class EdgeDetectionProcessor {
+}
