+++ app/src/main/java/com/project/snapsketch/domain/repository/ImageRepository.kt
@@ -1,0 +1,4 @@
+package com.project.snapsketch.domain.repository
+
+interface ImageRepository {
+}
