+++ app/src/main/java/com/project/snapsketch/domain/usecase/EdgeProcessImageUseCase.kt
@@ -1,0 +1,4 @@
+package com.project.snapsketch.domain.usecase
+
+class EdgeProcessImageUseCase {
+}
