+++ app/src/main/java/com/project/snapsketch/presentation/ui/album/AlbumViewModel.kt
@@ -1,0 +1,4 @@
+package com.project.snapsketch.presentation.ui.album
+
+class AlbumViewModel {
+}
