+++ app/src/main/java/com/project/snapsketch/presentation/ui/camera/CameraViewModel.kt
@@ -1,0 +1,4 @@
+package com.project.snapsketch.presentation.ui.camera
+
+class CameraViewModel {
+}
