+++ app/src/main/java/com/project/snapsketch/presentation/ui/coloring/ColoringViewModel.kt
@@ -1,0 +1,4 @@
+package com.project.snapsketch.presentation.ui.coloring
+
+class ColoringViewModel {
+}
