+++ app/src/main/java/com/project/snapsketch/presentation/ui/detecting/DetectingViewModel.kt
@@ -1,0 +1,4 @@
+package com.project.snapsketch.presentation.ui.detecting
+
+class DetectingViewModel {
+}
