+++ app/src/main/java/com/project/snapsketch/presentation/ui/home/HomeFragment.kt
@@ -7,13 +7,11 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
-import androidx.navigation.fragment.findNavController
 import com.project.snapsketch.R
 import com.project.snapsketch.databinding.FragmentHomeBinding
 import com.project.snapsketch.presentation.model.ImageModel
-import com.project.snapsketch.presentation.util.FileUtils
-import com.project.snapsketch.presentation.util.ToastMaker
+import com.project.snapsketch.data.local.FileUtils
+import com.project.snapsketch.presentation.utils.ToastMaker
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -21,7 +19,6 @@ class HomeFragment : Fragment() {
     private var _binding: FragmentHomeBinding? = null
     private val binding get() = _binding!!
 
-    private val viewmodel: HomeViewModel by viewModels()
     private val imageListAdapter: ImageListAdapter by lazy {
         ImageListAdapter(object : ImageListAdapter.ImageItemListener {
             override fun onItemClicked(item: ImageModel) {
@@ -66,8 +63,8 @@ class HomeFragment : Fragment() {
     }
 
     private fun goDetecting(uri: Uri) {
-        val action = HomeFragmentDirections.actionHomeFragmentToDetectingFragment(uri.toString())
-        findNavController().navigate(action)
+//        val action = HomeFragmentDirections.actionHomeFragmentToDetectingFragment(uri.toString())
+//        findNavController().navigate(action)
     }
 
     private fun setupImageList() {
