+++ app/src/main/java/com/project/snapsketch/presentation/ui/home/HomeViewModel.kt
@@ -1,6 +1,4 @@
 package com.project.snapsketch.presentation.ui.home
 
-import androidx.lifecycle.ViewModel
-
-class HomeViewModel : ViewModel() {
+class HomeViewModel {
 }
