--- app/src/main/java/com/project/snapsketch/presentation/ui/setting/SettingFragment.kt
@@ -1,19 +1,0 @@
-package com.project.snapsketch.presentation.ui.setting
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.project.snapsketch.R
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class SettingFragment : Fragment() {
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_setting, container, false)
-    }
-}
