--- app/src/main/res/menu/nav_menu.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:id="@+id/homeFragment"
-        android:enabled="true"
-        android:icon="@drawable/ic_home"
-        android:title="@string/main_home" />
-
-    <item
-        android:id="@+id/settingFragment"
-        android:enabled="true"
-        android:icon="@drawable/ic_setting"
-        android:title="@string/main_setting" />
-
-</menu>
