--- app/src/main/res/navigation/nav_graph.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/nav_graph"
-    app:startDestination="@id/homeFragment">
-
-    <fragment
-        android:id="@+id/homeFragment"
-        android:name="com.project.snapsketch.presentation.ui.home.HomeFragment"
-        android:label="fragment_home"
-        tools:layout="@layout/fragment_home">
-        <action
-            android:id="@+id/action_homeFragment_to_detectingFragment"
-            app:destination="@id/detectingFragment">
-            <argument
-                android:name="argsUriString"
-                android:defaultValue='""'
-                app:argType="string"
-                app:nullable="true" />
-        </action>
-    </fragment>
-
-    <fragment
-        android:id="@+id/detectingFragment"
-        android:name="com.project.snapsketch.presentation.ui.detecting.DetectingFragment"
-        android:label="fragment_detecting"
-        tools:layout="@layout/fragment_detecting">
-        <argument
-            android:name="argsUriString"
-            android:defaultValue='""'
-            app:argType="string"
-            app:nullable="true" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/settingFragment"
-        android:name="com.project.snapsketch.presentation.ui.setting.SettingFragment"
-        android:label="fragment_setting"
-        tools:layout="@layout/fragment_setting" />
-
-</navigation>
