+++ build.gradle.kts
@@ -9,7 +9,4 @@ buildscript {
     repositories {
         google()
     }
-    dependencies {
-        classpath(libs.androidx.navigation.safe.args.gradle.plugin)
-    }
 }
